[fix] DB 연결 수정

+++ eoullim-Project/src/main/java/org/example/eoullimback/review/Review.java
@@ -27,8 +27,9 @@ public class Review extends BaseTimeEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+
     @Column(name = "rating", nullable = false)
-    private Integer rating;
+    private Byte rating;
 
     @Column(name = "content", columnDefinition = "TEXT", nullable = false)
     private String content;
